refactor: move `Receipt` trait
