+++ crates/consensus/src/lib.rs
@@ -21,7 +21,7 @@ mod header;
 pub use header::{Header, EMPTY_OMMER_ROOT_HASH, EMPTY_ROOT_HASH};
 
 mod receipt;
-pub use receipt::{Receipt, ReceiptEnvelope, ReceiptWithBloom};
+pub use receipt::{Receipt, ReceiptEnvelope, ReceiptWithBloom, TxReceipt};
 
 mod transaction;
 pub use transaction::{
